Enemys

+++ Assets/_Complete-Game/Scripts/Wall.cs
@@ -8,7 +8,7 @@ namespace Completed
 		public AudioClip chopSound1;				//1 of 2 audio clips that play when the wall is attacked by the player.
 		public AudioClip chopSound2;				//2 of 2 audio clips that play when the wall is attacked by the player.
 		public Sprite dmgSprite;					//Alternate sprite to display after Wall has been attacked by player.
-		public int hp = 3;							//hit points for the wall.
+		public float hp = 3;							//hit points for the wall.
 		
 		
 		private SpriteRenderer spriteRenderer;		//Store a component reference to the attached SpriteRenderer.
@@ -22,7 +22,7 @@ namespace Completed
 		
 		
 		//DamageWall is called when the player attacks a wall.
-		public void DamageWall (int loss)
+		public void DamageWall (float loss)
 		{
 			//Call the RandomizeSfx function of SoundManager to play one of two chop sounds.
 			SoundManager.instance.RandomizeSfx (chopSound1, chopSound2);
